Initial checkin of relational algebra language.
The initial checkin should have sufficient functionality for use in
Caltech CS121.  It should support just about everything we do in class
slides and on relational algebra problems in assignments.  The "format
to HTML" functionality seems complete, but the interpretation support
is still very immature and not ready to use.

The build process is horrible - will fix that later.